refactor: enforce 50-line rule for AIBYOKWizard
